برنامه ای بنویسید که میزان مصرف ماشین در صد کیلومتر را و گنجایش باک ر…

+++ Examples3.java
@@ -8,9 +8,9 @@ public class Examples {
         float masafat = 0;
         Scanner input = new Scanner(System.in);
         System.out.print("لطفا میزان مصرف صد کیلومتر به لیتر را وارد کنید:");
-        masraf = input.nextInt();
+        masraf = input.nextFloat();
         System.out.print("لطفا گنجایش باک را وارد کنید:");
-        gonjayesh = input.nextInt();
+        gonjayesh = input.nextFloat();
         masafat = (gonjayesh / masraf) * 100;
         System.out.println("میزان مسافت قابل پیمایش با باک پر ماشین شما:" + masafat);
     }
